requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
mysql-connector-python==9.0.0
packaging==24.1
requests==2.32.3
SQLAlchemy==2.0.32
tabulate==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
